requirements.txt
customtkinter
tkinter
yt-dlp
ffmpeg-python
openai-whisper
googletrans==4.0.0-rc1
edge-tts
moviepy
pydub
libretranslate